Refactor state initialization into modular units
