+++ src/backend/src/state/initialization/finance.ts
@@ -1,0 +1,96 @@
+import type { BlueprintRepository } from '../../../data/blueprintRepository.js';
+import type { DeviceBlueprint, DevicePriceEntry } from '../../../data/schemas/index.js';
+import type {
+  EconomicsSettings,
+  FinanceState,
+  FinancialSummary,
+  LedgerEntry,
+  StructureBlueprint,
+} from '../models.js';
+import type { RngStream } from '../../lib/rng.js';
+import { generateId } from './common.js';
+
+const sumDeviceCapitalCosts = (
+  devices: DeviceBlueprint[],
+  priceLookup: (id: string) => DevicePriceEntry | undefined,
+): number => {
+  return devices.reduce((sum, device) => {
+    const price = priceLookup(device.id);
+    return sum + (price?.capitalExpenditure ?? 0);
+  }, 0);
+};
+
+export const createFinanceState = (
+  createdAt: string,
+  economics: EconomicsSettings,
+  blueprint: StructureBlueprint,
+  installedDevices: DeviceBlueprint[],
+  repository: BlueprintRepository,
+  idStream: RngStream,
+): FinanceState => {
+  const ledgerEntries: LedgerEntry[] = [];
+  const addEntry = (
+    entry: Omit<LedgerEntry, 'id' | 'tick' | 'timestamp'> & { tick?: number; timestamp?: string },
+  ) => {
+    ledgerEntries.push({
+      id: generateId(idStream, 'ledger'),
+      tick: entry.tick ?? 0,
+      timestamp: entry.timestamp ?? createdAt,
+      amount: entry.amount,
+      type: entry.type,
+      category: entry.category,
+      description: entry.description,
+    });
+  };
+
+  addEntry({
+    amount: economics.initialCapital,
+    type: 'income',
+    category: 'capital',
+    description: 'Initial capital injection',
+  });
+
+  if (blueprint.upfrontFee > 0) {
+    addEntry({
+      amount: -blueprint.upfrontFee,
+      type: 'expense',
+      category: 'structure',
+      description: `Lease upfront payment for ${blueprint.name}`,
+    });
+  }
+
+  const deviceCosts = sumDeviceCapitalCosts(installedDevices, (id) =>
+    repository.getDevicePrice(id),
+  );
+  if (deviceCosts > 0) {
+    addEntry({
+      amount: -deviceCosts,
+      type: 'expense',
+      category: 'device',
+      description: 'Initial device purchases',
+    });
+  }
+
+  const cashOnHand = economics.initialCapital - blueprint.upfrontFee - deviceCosts;
+  const totalExpenses = blueprint.upfrontFee + deviceCosts;
+
+  const summary: FinancialSummary = {
+    totalRevenue: economics.initialCapital,
+    totalExpenses,
+    totalPayroll: 0,
+    totalMaintenance: 0,
+    netIncome: economics.initialCapital - totalExpenses,
+    lastTickRevenue: economics.initialCapital,
+    lastTickExpenses: totalExpenses,
+  };
+
+  return {
+    cashOnHand,
+    reservedCash: 0,
+    outstandingLoans: [],
+    ledger: ledgerEntries,
+    summary,
+  };
+};
+
+export { sumDeviceCapitalCosts };
